+++ jgclark.RepeatExtensions/src/repeatExtensions.js
@@ -2,11 +2,12 @@
 //-----------------------------------------------------------------------
 // Repeat Extensions plugin for NotePlan
 // Jonathan Clark
-// last updated 9.3.2023 for v0.5.1+
+// last updated 14.3.2023 for v0.5.2
 //-----------------------------------------------------------------------
 
 import pluginJson from "../plugin.json"
 import {
+  calcOffsetDate,
   calcOffsetDateStr,
   getISOWeekString,
   isDailyNote,
@@ -16,6 +17,8 @@ import {
   RE_DATE_TIME,
   RE_DONE_DATE_TIME,
   RE_DONE_DATE_TIME_CAPTURES,
+  RE_SCHEDULED_DAILY_NOTE_LINK,
+  RE_SCHEDULED_WEEK_NOTE_LINK,
   RE_TIME, // find '12:23' with optional '[ ][AM|PM|am|pm]'
   unhyphenateString,
 } from '@helpers/dateTime'
@@ -33,7 +36,6 @@ import { showMessage } from '@helpers/userInput'
 //------------------------------------------------------------------
 // Regexes
 
-const RE_DUE_DATE_CAPTURE = `\\s+>(${RE_DATE})` // find ' >2021-02-23' and return just date part
 const RE_EXTENDED_REPEAT = `@repeat\\(${RE_DATE_INTERVAL}\\)` // find @repeat()
 const RE_EXTENDED_REPEAT_CAPTURE = `@repeat\\((.*?)\\)` // find @repeat() and return part inside brackets
 
@@ -53,12 +55,11 @@ export async function onEditorWillSave(): Promise<void> {
       // logDebug(pluginJson, `- previous version: ${String(noteReadOnly.versions[0].date)} [${previousContent}]`)
       // logDebug(pluginJson, `- new version: ${String(Date.now())} [${latestContent}]`)
 
-      // first check to see if this has been called in the last 1000ms: if so don't proceed, as this could be a double call.
-      // FIXME: As of 12.3.23 removing this check, and seeing what happens, following Discord discussion off #templating thread
-      // if (timeSinceLastEdit <= 2000) {
-      //   logDebug(pluginJson, `onEditorWillSave fired, but ignored, as it was called only ${String(timeSinceLastEdit)}ms after the last one`)
-      //   return
-      // }
+      // first check to see if this has been called in the last 2000ms: if so don't proceed, as this could be a double call.
+      if (timeSinceLastEdit <= 2000) {
+        logDebug(pluginJson, `onEditorWillSave fired, but ignored, as it was called only ${String(timeSinceLastEdit)}ms after the last one`)
+        return
+      }
 
       // Get changed ranges
       const ranges = NotePlan.stringDiff(previousContent, latestContent)
@@ -104,12 +105,15 @@ export async function onEditorWillSave(): Promise<void> {
  * Process any completed (or cancelled) tasks with my extended @repeat(..) tags,
  * and also remove the HH:MM portion of any @done(...) tasks.
  * When interval is of the form '+2w' it will duplicate the task for 2 weeks after the date is was completed.
- * When interval is of the form '2w' it will duplicate the task for 2 weeks after the date the task was last due.If this can't be determined, then default to the first option.
+ * When interval is of the form '2w' it will duplicate the task for 2 weeks after the date the task was last due. If this can't be determined, then default to the first option.
  * Valid intervals are [0-9][bdwmqy].
  * To work it relies on finding @done(YYYY-MM-DD HH:MM) tags that haven't yet been shortened to @done(YYYY-MM-DD).
  * It includes cancelled tasks as well; to remove a repeat entirely, remove the @repeat tag from the task in NotePlan.
+ * Note: The new repeat date is by default scheduled to a day (>YYYY-MM-DD).
+ * But if the scheduled date is a week date (YYYY-Wnn), or the repeat is in a weekly note, then the new repeat date will be a scheduled week link (>YYYY-Wnn).
+ * Note: Runs on the currently open note (using Editor.* funcs)
  * Note: Could add a 'Newer' mode of operation according to #351.
- * Note: Mostly operates on the currently open note (using Editor.* funcs)
+ * TEST: fails to appendTodo to note with same stem?
  * @author @jgclark
  * @param {CoreNoteFields} noteIn?
  */
@@ -177,111 +181,113 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
         p.content = updatedLine
         // Send the update to the Editor
         Editor.updateParagraph(p)
-        logDebug('repeats', `- updated Paragraph ${p.lineIndex}`)
+        logDebug('repeats', `- updated Para ${p.lineIndex} -> <${updatedLine}>`)
 
         // Test if this is one of my special extended repeats
         if (updatedLine.match(RE_EXTENDED_REPEAT)) {
           repeatCount++
-          let newRepeatDate = ''
+          let newRepeatDateStr = ''
+          let newRepeatDate: Date
           let outputLine = ''
+
           // get repeat to apply
           reReturnArray = updatedLine.match(RE_EXTENDED_REPEAT_CAPTURE) ?? []
           let dateIntervalString: string = (reReturnArray.length > 0) ? reReturnArray[1] : ''
           logDebug('repeats', `- Found EXTENDED @repeat syntax: '${dateIntervalString}'`)
 
+          // decide style of new date: daily or weekly link
+          let timeframe = 'day'
+          if (updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK) || isWeeklyNote(note)) {
+            timeframe = 'week'
+          }
+          // TODO: use this in >day case in weekly note
+          logDebug('repeats', `- timeframe: ${timeframe} from ${String(updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK))} / ${String(isWeeklyNote(note))}`)
+
           if (dateIntervalString[0].startsWith('+')) {
-            // New repeat date = completed date + interval
+            // New repeat date = completed date (of form YYYY-MM-DD) + interval
             dateIntervalString = dateIntervalString.substring(
               1,
               dateIntervalString.length,
             )
-            newRepeatDate = calcOffsetDateStr(completedDate, dateIntervalString)
-            logDebug('repeats', `- Adding from completed date -> ${newRepeatDate}`)
+            newRepeatDateStr = calcOffsetDateStr(completedDate, dateIntervalString)
+            newRepeatDate = calcOffsetDate(completedDate, dateIntervalString) ?? new Date()
+            logDebug('repeats', `- Adding from completed date -> ${newRepeatDateStr}, ${String(newRepeatDate)}`)
             // Remove any >date
             // Note: ' >'+RE_DUE_DATE, but can't get regex to work with variables like this
             updatedLine = updatedLine.replace(/\s+>\d{4}-[01]\d{1}-\d{2}/, '')
+            // TODO: or week version?
             logDebug('repeats', `- updatedLine: ${updatedLine}`)
 
           } else {
             // New repeat date = due date + interval
-            // look for the due date(>YYYY-MM-DD)
+            // look for the due date (>YYYY-MM-DD)
             let dueDate = ''
-            const resArray = updatedLine.match(RE_DUE_DATE_CAPTURE) ?? []
-            // logDebug('repeats', resArray.length)
-            if (resArray[1] != null) {
-              logDebug('repeats', `  match => ${resArray[1]}`)
-              dueDate = resArray[1]
+            const dueDateArray = RE_SCHEDULED_DAILY_NOTE_LINK.test(updatedLine)
+              ? updatedLine.match(RE_SCHEDULED_DAILY_NOTE_LINK)
+              : RE_SCHEDULED_WEEK_NOTE_LINK.test(updatedLine)
+                ? updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK)
+                : []
+            logDebug('repeats', `- dueDateArray: ${String(dueDateArray)}`)
+            if (dueDateArray && dueDateArray[0] != null) {
+              dueDate = dueDateArray[0].split('>')[1]
+              // logDebug('repeats', `  due date match => ${dueDate}`)
               // need to remove the old due date
-              updatedLine = updatedLine.replace(`>${dueDate}`, '')
-              // logDebug('repeats', updatedLine);
+              updatedLine = updatedLine.replace(` >${dueDate}`, '')
             } else {
               // but if there is no due date then treat that as today
               dueDate = completedDate
               logDebug('repeats', `- no match => use completed date ${dueDate}`)
             }
-            newRepeatDate = calcOffsetDateStr(dueDate, dateIntervalString)
-            logDebug('repeats', `- Adding from due date -> ${newRepeatDate}`)
+            newRepeatDateStr = calcOffsetDateStr(dueDate, dateIntervalString)
+            newRepeatDate = calcOffsetDate(dueDate, dateIntervalString) ?? new Date()
+            logDebug('repeats', `- adding from due date -> ${newRepeatDateStr}, ${String(newRepeatDate)}`)
           }
 
           outputLine = updatedLine.replace(/@done\(.*\)/, '').trim()
 
           // Create and add the new repeat line
           if (type === 'Notes') {
-            // ...either in same project note
-            outputLine += ` >${newRepeatDate}`
-            logDebug(pluginJson, `- outputLine: ${outputLine}`)
+            // ...either in same project note, including new scheduled date
+            outputLine += ` >` + newRepeatDateStr
+            logDebug(pluginJson, `- outputLine: <${outputLine}>`)
             await Editor.insertParagraphBeforeParagraph(outputLine, p, 'open')
             logDebug(pluginJson, `- Inserted new para after line ${p.lineIndex}`)
           }
           else {
-            // ... or in the future Calendar note (prepend)
-            let futureNote
-            let newRepeatDate = ''
-            if (isDailyNote(note)) {
-              // Get YYYY-MM-DD style date
-              newRepeatDate = unhyphenateString(newRepeatDate)
-            }
-            else if (isWeeklyNote(note)) {
-              // Get YYYY-Wnn style date
-              // older version, but doesn't align with NP user week-start setting
-              // newRepeatDate = getISOWeekString(newRepeatDate)
-              // newer version, using helper that takes week-start into account
-              // $FlowFixMe[incompatible-type]
-              const weekInfo: NotePlanWeekInfo = getNPWeekData(newRepeatDate)
-              newRepeatDate = weekInfo.weekString
-            }
-            logInfo(pluginJson, `- repeat -> ${newRepeatDate}`)
+            // ... or in the future Calendar note
+            // let futureNote
+            // let newRepeatDateStr = ''
+            // if (isDailyNote(note)) {
+            //   // Get YYYY-MM-DD style date
+            //   newRepeatDateStr = unhyphenateString(newRepeatDateStr)
+            // }
+            // else if (isWeeklyNote(note)) {
+            //   // Get YYYY-Wnn style date
+            //   // older version, but doesn't align with NP user week-start setting
+            //   // newRepeatDateStr = getISOWeekString(newRepeatDateStr)
+            //   // newer version, using helper that takes week-start into account
+            //   // $FlowFixMe[incompatible-type]
+            //   const weekInfo: NotePlanWeekInfo = getNPWeekData(newRepeatDateStr)
+            //   newRepeatDateStr = weekInfo.weekString
+            // }
+            logInfo(pluginJson, `- repeat -> ${newRepeatDateStr}`)
 
-            futureNote = await DataStore.calendarNoteByDateString(newRepeatDate)
+            // Grr: can't use this next one, as it doesn't support undefined notes
+            // let futureNote = await DataStore.calendarNoteByDateString(newRepeatDateStr)
+            let futureNote = await DataStore.calendarNoteByDate(newRepeatDate, timeframe)
             if (futureNote != null) {
+              // Add todo to future note
               logDebug(pluginJson, futureNote.filename)
               await futureNote.appendTodo(outputLine)
-              logDebug(pluginJson, `- Appended new repeat in calendar note '${newRepeatDate}'`)
+              logDebug(pluginJson, `- Appended new repeat in calendar note ${displayTitle(futureNote)}`)
             } else {
               // After a fix to future calendar note creation in r635, we shouldn't get here.
               // But just in case, we'll insert new repeat into the open note
-              outputLine += ` >${newRepeatDate}`
+              outputLine += ` >${newRepeatDateStr}`
               logDebug('repeats', `- outputLine: ${outputLine}`)
               await Editor.insertParagraphBeforeParagraph(outputLine, p, 'open')
-              logDebug('repeats', `- Inserted new para after line ${p.lineIndex}`)
+              logDebug('repeats', `- Couldn't get futureNote, so instead inserted new para after line ${p.lineIndex} in original note`)
             }
-            // TODO: after a big merge not sure if this is needed.
-            // else {
-            //   // ... or in the future daily note (prepend)
-            //   const newRepeatDateShorter = unhyphenateString(newRepeatDate)
-            //   const newDailyNote = await DataStore.calendarNoteByDateString(newRepeatDateShorter)
-            //   if (newDailyNote?.title != null) {
-            //     logDebug('repeats', newDailyNote.filename)
-            //     await newDailyNote.appendTodo(outputLine)
-            //     logDebug('repeats', `  Inserted new repeat in daily note ${newRepeatDateShorter}`)
-            //   } else {
-            //     // After a fix to future calendar note creation in r635, we shouldn't get here.
-            //     // But just in case, we'll create new repeat in today's daily note
-            //     outputLine += ` >${newRepeatDate}`
-            //     logDebug('repeats', `- outputLine: ${outputLine}`)
-
-            //     await Editor.insertParagraphAfterParagraph(outputLine, p, 'open')
-            //     logWarn('repeats', 'Inserted new repeat in original daily note')
           }
         }
       }